+++ source/benchmarks/multitile_memory_benchmark/implementations/l0/usm_shared_migrate_cpu_l0.cpp
@@ -58,7 +58,7 @@ static TestResult run(const UsmSharedMigrateCpuArguments &arguments, Statistics 
 
     // Warmup
     ASSERT_ZE_RESULT_SUCCESS(zeCommandQueueExecuteCommandLists(levelzero.commandQueue, 1, &cmdList, nullptr));
-    ASSERT_ZE_RESULT_SUCCESS(zeCommandQueueSynchronize(levelzero.commandQueue, std::numeric_limits<uint32_t>::max()));
+    ASSERT_ZE_RESULT_SUCCESS(zeCommandQueueSynchronize(levelzero.commandQueue, std::numeric_limits<uint64_t>::max()));
     for (auto elementIndex = 0u; elementIndex < elementsCount; elementIndex++) {
         buffer[elementIndex] = 0;
     }
@@ -66,7 +66,7 @@ static TestResult run(const UsmSharedMigrateCpuArguments &arguments, Statistics 
     // Benchmark
     for (auto i = 0; i < arguments.iterations; i++) {
         ASSERT_ZE_RESULT_SUCCESS(zeCommandQueueExecuteCommandLists(levelzero.commandQueue, 1, &cmdList, nullptr));
-        ASSERT_ZE_RESULT_SUCCESS(zeCommandQueueSynchronize(levelzero.commandQueue, std::numeric_limits<uint32_t>::max()));
+        ASSERT_ZE_RESULT_SUCCESS(zeCommandQueueSynchronize(levelzero.commandQueue, std::numeric_limits<uint64_t>::max()));
 
         timer.measureStart();
         if (arguments.accessAllBytes) {
